backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
click==8.1.7
Django==4.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-rq==2.10.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
numpy==2.1.1
pandas==2.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
redis==5.1.0
rq==1.16.2
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
